refactor(sdk): move ITeam Interface to a better location
