+++ plugin/src/shared/sc/plugin2020/Piece.kt
@@ -2,7 +2,6 @@ package sc.plugin2020
 
 import com.thoughtworks.xstream.annotations.XStreamAlias
 import com.thoughtworks.xstream.annotations.XStreamAsAttribute
-import sc.shared.ITeam
 
 @XStreamAlias(value = "piece")
 data class Piece(
