+++ plugin/src/shared/sc/plugin2020/Team.kt
@@ -1,6 +1,6 @@
 package sc.plugin2020
 
-import sc.shared.ITeam
+import sc.api.plugins.ITeam
 
 enum class Team(override val index: Int, val displayName: String): ITeam {
     RED(0, "Rot") {
