+++ plugin/src/shared/sc/plugin2021/Team.kt
@@ -1,6 +1,6 @@
 package sc.plugin2021
 
-import sc.shared.ITeam
+import sc.api.plugins.ITeam
 
 enum class Team(override val index: Int, val colors: List<Color>): ITeam<Team> {
     ONE(0, listOf(Color.BLUE, Color.YELLOW)) {
