+++ plugin/src/test/sc/plugin2020/CloneTest.kt
@@ -4,7 +4,7 @@ import io.kotlintest.matchers.types.shouldNotBeSameInstanceAs
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.StringSpec
 import sc.framework.plugins.Player
-import sc.shared.ITeam
+import sc.api.plugins.ITeam
 
 class CloneTest: StringSpec({
     "clone Player" {
