+++ server/test/sc/server/helpers/TestTeam.kt
@@ -1,6 +1,6 @@
 package sc.server.helpers
 
-import sc.shared.ITeam
+import sc.api.plugins.ITeam
 
 enum class TestTeam(override val index: Int, val displayName: String): ITeam<TestTeam> {
     RED(0, "Rot") {
