+++ server/test/sc/server/plugins/TestPlayer.java
@@ -3,7 +3,7 @@ package sc.server.plugins;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sc.framework.plugins.Player;
-import sc.shared.ITeam;
+import sc.api.plugins.ITeam;
 
 public class TestPlayer extends Player {
   private static Logger logger = LoggerFactory.getLogger(TestPlayer.class);
