+++ socha-sdk/src/framework/sc/shared/WelcomeMessage.kt
@@ -2,6 +2,7 @@ package sc.shared
 
 import com.thoughtworks.xstream.annotations.XStreamAlias
 import com.thoughtworks.xstream.annotations.XStreamAsAttribute
+import sc.api.plugins.ITeam
 import sc.protocol.responses.ProtocolMessage
 
 /** Nachricht, die zu Beginn eines Spiels an einen Client geschickt wird, um ihm seine Spielerfarbe mitzuteilen.  */
