+++ socha-sdk/src/framework/sc/shared/WinCondition.java
@@ -1,5 +1,7 @@
 package sc.shared;
 
+import sc.api.plugins.ITeam;
+
 public class WinCondition implements Cloneable {
 
   /** Farbe des Gewinners */
